+++ src/style/style.h
@@ -54,8 +54,8 @@ class Style : public QAbstractTableModel, public Meta {
  private:
   bool changed{false}; /**< @brief True if changed but not saved */
 
-  std::array<bool, static_cast<int>(Water::Value::Size)> limited{};
-  std::array<std::array<double, static_cast<int>(Limit::Size)>, static_cast<int>(Water::Value::Size)> limits{};
+  std::array<bool, static_cast<int>(Water::Value::ExtendedSize)> limited{};
+  std::array<std::array<double, static_cast<int>(Limit::Size)>, static_cast<int>(Water::Value::ExtendedSize)> limits{};
   // NOLINTNEXTLINE(cert-err58-cpp): it is very unlikely and program will not run without this strings
   static const inline std::array<QString, static_cast<int>(Limit::Size)> strJsonKey{"Min", "Target", "Max"};
 };